Added support for assets in the Converter and removed pandoc and markdown support
Assets are now gathered in RST directives and copied to the destination location. Because
this new approach is not present in the markdown code and we want to remove the
dependency to pandoc we have removed the Markdown writers and will re-add them
once the RST and basic design is fully implemented.